fixed the paramters types

+++ ft_bzero.c
@@ -6,13 +6,14 @@
 /*   By: zaalrafa <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/21 00:14:37 by zaalrafa          #+#    #+#             */
-/*   Updated: 2025/08/21 00:19:09 by zaalrafa         ###   ########.fr       */
+/*   Updated: 2025/08/21 00:50:32 by zaalrafa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+#include "libft.h"
 
-void	ft_bzero(char *src, int n)
+void	ft_bzero(char *src, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i <= n)
